Refactor Google OAuth adapter boundaries
